Fix crash when rapidly scrolling
If you scroll too fast, this can come back with `nil` data. If that
happens, `UIImage(data:)` will fail and return `nil` even though the
type says that shouldn't be possible. This is going to be the behavior
with Swift 1.1 anyway, so this is a bit of futureproofing as well.