Rename view models
We don't need a cell specific view model. It seems like it's more useful
to have a view model that represents a specific user. That way we can do
any transformations we need to do in one place. This essentially turns
the view model into a facade for use in displaying the model.

To reduce confusion, I'm also renaming EmployeesViewModel to
EmployeeListViewModel, which is actually a little more accurate about
what it's modeling.